+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix overlapping text in boss selector
+
 ## [0.12.0] - 2026-04-15
 
 - Fixed crash on opening inventory on latest patch